Fixing ring not working when device is muted
When the ring operation is added to the device and if the devices volume is muted, the ring sound is not audible. Although the ringer volume is at maximum, since AVAudioPlayer is used for audio playing, the ringer volume has no effect. Since MPMusicPlayerController.volume is depricated, we have to use the work around, where we create a volume slider and adjust the volume.

Fixed wso2/product-iots#1668